Rewriters: Hide internal details of rewriters
DivModRewriter is exposed to the outside world, but it was broken,
because it was including "OsmtInternalException.h" which is not suppose
to be exposed and which we do not install. Thus, trying to include
"DivModRewriter.h" in another application results in a compilation
error.

We fix this problem by hiding the existing rewriters and exposing to the
outside world only API methods to use this functionality.