Tests: Remove outdated test
The file test_LIACutSolver contains only one test, which is currently
not testing any useful functionality.
It seems like this test originated in 2019 as part of preparation for a
new type of LIA solver. However, there has been a lot of changes in the
architecture of the arithmetic solver since then and this "LIACutSolver"
was removed a long time ago.

The test attempts to create an instance of simplex, however, there are
no bounds asserted, nor any basic variables are marked as out of bounds,
so the check does not actually do any work.
Simplex is not prepared to be test in this way anyway, so this test
should be removed.