Core: Fix split clause handling
So-called split clauses from array solver can contain many literals.
It can happen that a split clause is satisfied,
but the first two literals are falsified.
The code for processing new split clauses would previosuly take the
clause as is and try to attach it to the SAT solver. However,
the method attachClause has a check that at the time of attaching, the
first two literals cannot be both falsified.
We already had a check in place for the situation where the two first
literals are falsified, but there is at least one undecided literal.
However, it may happen that there is no undecided literal (all others
are satisfied). In this case, we also need to swap one of the satisfied
literals with the first, falsified, literal.