Substitutions: Add detection of conflict in arithmetic substitutions
Previously, the test I am adding here would run into an assertion,
because it would try to create a substitution `x -> 1/3` in the integer
logic.
To prevent that, we can detect such cases and abort the computation of
substitutions early. We have to pass along the fact that the formula
from which we are computing substitutions is unsatisfiable.