Fix assertion violation in split clauses
Previously, it could have happened that both the arithmetic solver and
the array solver would produce a new lemma for the SAT solver.
In our current implementation we prefer arithmetic lemmas, so if there
is a new arithmetic lemma, we ignore the ones from array solver.
However, we would previously not clear the stored lemmas in the array
solver. So if the solver would backtrack due to the arithmetic lemma and
change the assignment, the array lemma might be outdated (e.g., already
satisfied). If only a single satisfied lemma is given to the code
handling newly introduced lemmas, that code runs into an assertion.
The proposed solution is to clear the lemmas in the array solver when
the SAT solver backtracks.